Convert tests to TS, to ensure types are working
